+++ src/main/java/com/imsweb/algorithms/Algorithms.java
@@ -145,6 +145,8 @@ public final class Algorithms {
     public static final String FIELD_AYA_SITE_RECODE_2020 = "ayaSiteRecode2020";
     public static final String FIELD_ICCC = "iccc";
     public static final String FIELD_ICCC_MAJOR_CATEGORY = "icccMajorCategory";
+    public static final String FIELD_ICCC_EXT = "icccExtended";
+    public static final String FIELD_ICCC_EXT_MAJOR_CATEGORY = "icccExtendedMajorCategory";
     public static final String FIELD_IARC_MP_INDICATOR = "iarcMpIndicator";
     public static final String FIELD_IARC_MP_SITE_GROUP = "iarcMpSiteGroup";
     public static final String FIELD_IARC_MP_HIST_GROUP = "iarcMpHistGroup";
@@ -268,8 +270,10 @@ public final class Algorithms {
             addField(AlgorithmField.of(FIELD_SEER_BEHAV_RECODE, null, 1, "SEER Site Behavior Recode ICD-O-3", "SEER Behav", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_AYA_SITE_RECODE_2008, null, 2, "AYA Site Recode 2008", "AYA 08", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_AYA_SITE_RECODE_2020, null, 3, "AYA Site Recode 2020", "AYA 20", DATA_LEVEL_TUMOR));
-            addField(AlgorithmField.of(FIELD_ICCC, null, 3, "Int Classification of Childhood Cancer (ICCC)", "ICCC", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_ICCC, null, 3, "Int Class of Childhood Cancer (ICCC)", "ICCC", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_ICCC_MAJOR_CATEGORY, null, 2, "ICCC Major Category", "ICCC Cat", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_ICCC_EXT, null, 3, "Int Class of Childhood Cancer (ICCC) Extended Code", "ICCC Ext", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_ICCC_EXT_MAJOR_CATEGORY, null, 2, "ICCC Extended Major Category", "ICCC Ext Cat", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_IARC_MP_INDICATOR, null, 1, "IARC Multiple Primary Indicator", "IARC MP", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_IARC_MP_SITE_GROUP, null, 3, "IARC Multiple Primary Site Group", "IARC MP Site Gr", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_IARC_MP_HIST_GROUP, null, 2, "IARC Multiple Primary Histology Group", "IARC MP Hist Grp", DATA_LEVEL_TUMOR));